refactor callback manager
