+++ cms-adapter/src/Rsbc.Dmf.CaseManagement/Manager/Callback/ICallbackManager.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Threading.Tasks;
+
+namespace Rsbc.Dmf.CaseManagement
+{
+    public interface ICallbackManager
+    {
+        Task<IEnumerable<Callback>> GetDriverCallbacks(Guid driverId);
+    }
+}
