+++ cms-adapter/src/Rsbc.Dmf.CaseManagement/Manager/CaseManager.cs
@@ -140,28 +140,6 @@ namespace Rsbc.Dmf.CaseManagement
         public string DriverLicenseNumber { get; set; }
     }
 
-    // TODO sync with GRPC models?
-    public enum CallbackTopic
-    {
-        View = 0,
-        Upload = 1
-    }
-
-    public enum CallbackCallStatus
-    {
-        Open = 0,
-        Closed = 1
-    }
-
-    public class Callback
-    {
-        public Guid Id { get; set; }
-        public DateTimeOffset RequestCallback { get; set; }
-        public CallbackTopic Topic { get; set; }
-        public CallbackCallStatus CallStatus { get; set; }
-        public DateTimeOffset Closed { get; set; }
-    }
-
     public class Provider
     {
         public string Id { get; set; }
